requirements.txt
chardet==5.0.0
click==8.1.3
colorama==0.4.5
dateparser==1.1.1
decorator==5.1.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
invoice2data==0.3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
PyJWT==2.4.0
PyPDF2==2.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.39
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.4
validators==0.20.0
Werkzeug==2.1.2